// client/src/types/listing.ts
export interface IListing {
  lid: number;
  bid: number; 
  uid: number;
  dateListed: string;
  price: number;
  bedCount: number;
  bathCount: number;
  squareFootage: number; 
  listingType: number;
  streetNumber: number; 
  streetName: string; 
  city: string; 
  state: string; 
  zipCode: number;
}